Update devcontainer.json and molecule.yml configurations
This commit updates the devcontainer.json file to include additional extensions for Visual Studio Code, such as EditorConfig, ms-python.autopep8, ms-python.flake8, redhat.ansible, and redhat.vscode-yaml. It also sets the default formatter for Python files to ms-python.autopep8 and enables formatting on save. Additionally, the ansible.python.interpreterPath is set to /usr/local/bin/python and the python.formatting.provider is set to none.

The molecule.yml file is also updated to use different platform images for testing. The amazonlinux-2023 and alpine-3.20 platforms are added, and the existing debian-10 platform is replaced with debian-10. The oraclelinux-9 and rockylinux-9 platforms are also added.